# Revise `sample_z`