--- a/benches/benchmarks.rs
+++ b/benches/benchmarks.rs
@@ -13,7 +13,15 @@ use criterion::criterion_main;
 pub mod cholesky_decomposition;
 pub mod classic_crypto;
 pub mod integer;
+pub mod sample_z;
 pub mod sampling;
 pub mod solve;
 
-criterion_main! {integer::benches, classic_crypto::benches, sampling::benches, solve::benches, cholesky_decomposition::benches}
+criterion_main! {
+    integer::benches,
+    classic_crypto::benches,
+    sampling::benches,
+    solve::benches,
+    cholesky_decomposition::benches,
+    sample_z::benches
+}