--- a/src/integer/z/hash.rs
+++ b/src/integer/z/hash.rs
@@ -6,16 +6,21 @@
 // the terms of the Mozilla Public License Version 2.0 as published by the
 // Mozilla Foundation. See <https://mozilla.org/en-US/MPL/2.0/>.
 
-//! `Z` is a type for integers with arbitrary length.
-//! This implementation uses the [FLINT](https://flintlib.org/) library.
+//! Implementation of the [`Hash`] trait for [`Z`].
 
 use super::Z;
 use std::hash::Hash;
 
+// MODULUS is 2^62 - 57, which is the largest prime below 2^62.
+// We choose it prime to reduce the risk of accidentally introducing
+// a hash collision by multiplying values by a prime factor of the given modulus.
+static MODULUS: i64 = 4611686018427387847;
+
 impl Hash for Z {
-    /// This function uses the hash of [`String`] with the decimal representation of `self`.
+    /// Calculates the value of `self` mod 2^62 - 57 and then uses
+    /// the inbuilt hash function for [`i64`].
     fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        let string = self.to_string();
-        string.hash(state);
+        let modulo_value = self.modulo(MODULUS).value.0;
+        modulo_value.hash(state);
     }
 }