--- a/src/integer_mod_q/mat_zq/sample/discrete_gauss.rs
+++ b/src/integer_mod_q/mat_zq/sample/discrete_gauss.rs
@@ -14,7 +14,9 @@ use crate::{
     integer_mod_q::{MatZq, Modulus},
     rational::{MatQ, Q},
     traits::{GetNumColumns, GetNumRows, SetEntry},
-    utils::sample::discrete_gauss::{sample_d, sample_d_precomputed_gso, sample_z},
+    utils::sample::discrete_gauss::{
+        sample_d, sample_d_precomputed_gso, DiscreteGaussianIntegerSampler,
+    },
 };
 use std::fmt::Display;
 
@@ -43,7 +45,7 @@ impl MatZq {
     ///
     /// # Errors and Failures
     /// - Returns a [`MathError`] of type [`InvalidIntegerInput`](MathError::InvalidIntegerInput)
-    ///     if `n <= 1` or `s <= 0`.
+    ///     if `n <= 1` or `s <= 0` or `s * log_2(n) < 1` or `s * log_2(n) < 1`.
     ///
     /// # Panics ...
     /// - if the provided number of rows and columns or the modulus are not suited to create a matrix.
@@ -62,9 +64,11 @@ impl MatZq {
         let s: Q = s.into();
         let mut out = Self::new(num_rows, num_cols, modulus);
 
+        let mut dgis = DiscreteGaussianIntegerSampler::init(&n, &center, &s)?;
+
         for row in 0..out.get_num_rows() {
             for col in 0..out.get_num_columns() {
-                let sample = sample_z(&n, &center, &s)?;
+                let sample = dgis.sample_z();
                 out.set_entry(row, col, sample).unwrap();
             }
         }