--- a/src/integer_mod_q/poly_over_zq/sample/discrete_gauss.rs
+++ b/src/integer_mod_q/poly_over_zq/sample/discrete_gauss.rs
@@ -15,7 +15,7 @@ use crate::{
     integer_mod_q::{Modulus, PolyOverZq},
     rational::Q,
     traits::SetCoefficient,
-    utils::{index::evaluate_index, sample::discrete_gauss::sample_z},
+    utils::{index::evaluate_index, sample::discrete_gauss::DiscreteGaussianIntegerSampler},
 };
 use std::fmt::Display;
 
@@ -45,7 +45,7 @@ impl PolyOverZq {
     ///
     /// # Errors and Failures
     /// - Returns a [`MathError`] of type [`InvalidIntegerInput`](MathError::InvalidIntegerInput)
-    ///     if `n <= 1` or `s <= 0`.
+    ///     if `n <= 1` or `s <= 0` or `s * log_2(n) < 1`.
     ///
     /// # Panics ...
     /// - if `max_degree` is negative, or does not fit into an [`i64`].
@@ -65,8 +65,10 @@ impl PolyOverZq {
         let s = s.into();
         let mut poly = PolyOverZq::from(&modulus);
 
+        let mut dgis = DiscreteGaussianIntegerSampler::init(&n, &center, &s)?;
+
         for index in 0..=max_degree {
-            let sample = sample_z(&n, &center, &s)?;
+            let sample = dgis.sample_z();
             poly.set_coeff(index, &sample)?;
         }
         Ok(poly)