**Description**

This PR implements `DiscreteGaussianIntegerSampler` and sample_z to
- [x] store evaluations of the Gaussian function efficiently (requires implementation of `Hash` for `Z`),
- [x] use f64::exp instead of Q::exp (roughly twice as fast - though there is no specification given for its precision),
- [x] corrects the interval that sample_z samples uniformly from, which was one entry too small,
- [x] benchmarks for sample_z.

Overall, the new implementation reduces discrete Gaussian sampling for a single sample by roughly 10 µs for the evaluated parameters (from 28 µs to 17 µs for a wide distribution and 18 µs to 9 µs for a narrow distribution).
If we sample multiple samples from the same distribution (usually done to sample the G-trapdoor R or preimages for G), then the speedup is just shy of 90% for 10,000 samples (280 ms -> 32 ms for a wide and 180 ms -> 19 ms for a narrow distribution).

**Testing**
- [x] I added basic working examples (possibly in doc-comment)
- [x] I added tests for large (pointer representation) values
- [x] I triggered all possible errors in my test in every possible way
- [x] I included tests for all reasonable edge cases

**Checklist:**
- [x] I have performed a self-review of my own code
  - [x] The code provides good readability and maintainability s.t. it fulfills best practices like talking code, modularity, ...
    - [x] The chosen implementation is not more complex than it has to be
  - [x] My code should work as intended and no side effects occur (e.g. memory leaks)
  - [x] The doc comments fit our style guide
  - [x] I have credited related sources if needed